% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robets.R
\name{coef.robets}
\alias{coef.robets}
\title{Coef robets model}
\usage{
\method{coef}{robets}(object, ...)
}
\arguments{
\item{object}{An object of class \code{robets}.}
}
\description{
Coef robets model
}
\examples{
model = robets(USAccDeaths)
coef(model)
}

