% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.robets}
\alias{plot.robets}
\title{Plot robets model}
\usage{
\method{plot}{robets}(x, ...)
}
\arguments{
\item{x}{An object of class \code{robets}.}

\item{...}{Other plotting parameters.}
}
\description{
Plot robets model
}
\examples{
model <- robets(nottem)
plot(model)
}
\seealso{
\code{\link{plotOutliers}, \link{plot.ets}}
}
