% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robets.R
\name{plotOutliers}
\alias{plotOutliers}
\title{Plot outliers detected by robets model}
\usage{
plotOutliers(object, xlab = "", ylab = "", type = "l", ...)
}
\arguments{
\item{object}{An object of class \code{robets}.}

\item{...}{Other plotting parameters.}
}
\description{
Plot outliers detected by robets model
}
\examples{
model = robets(USAccDeaths)
plotOutliers(model)
}
\seealso{
\code{\link{plot.robets}}
}

