% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.robets}
\alias{print.robets}
\title{Print robets model}
\usage{
\method{print}{robets}(x, ...)
}
\arguments{
\item{x}{An object of class \code{robets}.}

\item{...}{Other undocumented arguments.}
}
\description{
Print robets model
}
\examples{
model <- robets(nottem)
print(model)
}
