% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robets.R
\name{summary.robets}
\alias{summary.robets}
\title{Summary robets model}
\usage{
\method{summary}{robets}(object, ...)
}
\arguments{
\item{object}{An object of class \code{robets}.}

\item{...}{Other undocumented arguments.}
}
\description{
Summary robets model
}

