% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robets.R
\name{tau2}
\alias{tau2}
\title{Compute the tau2 estimator of scale}
\usage{
tau2(x)
}
\arguments{
\item{x}{A vector of residuals.}
}
\value{
The tau2 estimate of scale.
}
\description{
The tau2-estimator is a robust measure of the scale. The exact formula of the estimator is in Gelper et al. (2010), equation 16.
}
\examples{
set.seed(100)
e <- 10*rnorm(100)
mse <- mean(e^2) 
tse <- tau2(e) 
}
\references{
Gelper S., Fried R. and Croux C. (2010) "Robust Forecasting with Exponential and Holt-Winters Smoothing".\emph{Journal of Forecasting}, \strong{29}, 285-300.
}

