% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.robets}
\alias{coef.robets}
\title{Coef robets model}
\usage{
\method{coef}{robets}(object, ...)
}
\arguments{
\item{object}{An object of class \code{robets}.}

\item{...}{Other undocumented arguments.}
}
\description{
Coef robets model
}
\examples{
model <- robets(nottem)
coef(model)
}
