% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOutliers.R
\name{plotOutliers}
\alias{plotOutliers}
\title{Plot outliers detected by robets model}
\usage{
plotOutliers(object, xlab = "", ylab = "", type = "l", ...)
}
\arguments{
\item{object}{An object of class \code{robets}.}

\item{xlab}{Label of the x-axis.}

\item{ylab}{Label of the y-asix.}

\item{type}{Character indicating the type of plot, just as in \code{plot}.}

\item{...}{Other plotting parameters.}
}
\description{
Plot outliers detected by robets model
}
\examples{
model <- robets(nottem)
plotOutliers(model)
}
\seealso{
\code{\link{plot.robets}}
}
