% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.robets}
\alias{summary.robets}
\title{Summary robets model}
\usage{
\method{summary}{robets}(object, ...)
}
\arguments{
\item{object}{An object of class \code{robets}.}

\item{...}{Other undocumented arguments.}
}
\value{
A number of training set error measures: ME (mean error), RMSE (root mean squared error), MAE (mean absolute error), MPE (mean percentage error), MAPE (mean absolute percentage error), MedianE (median error), RTSE (root tau squared error), RTSPE (root tau squared percentage error).
}
\description{
Summary robets model
}
\examples{
model <- robets(nottem)
summary(model)
}
